fix: move files to respective directories
