+++ apps/drec-api/test/user.e2e-spec.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-unused-expressions */
 import { HttpStatus, INestApplication } from '@nestjs/common';
 import request from 'supertest';
-import { DatabaseService } from '../src/pods/origin-backend-utils/database.service';
+import { DatabaseService } from '../src/pods/utils/origin-backend-utils/database.service';
 
 import { bootstrapTestInstance } from './drec-api';
 import { UserService } from '../src/pods/user/user.service';
